+++ flutter_google_street_view/example/lib/demo/street_view_panorama_init.dart
@@ -1,7 +1,19 @@
+import 'dart:io';
+import 'dart:typed_data';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_google_street_view/flutter_google_street_view.dart';
+import 'package:flutter_google_street_view_example/const/const.dart';
 
-class StreetViewPanoramaInitDemo extends StatelessWidget {
+class StreetViewPanoramaInitDemo extends StatefulWidget {
+  @override
+  State<StatefulWidget> createState() => _StreetViewPanoramaInitDemoState();
+}
+
+class _StreetViewPanoramaInitDemoState
+    extends State<StreetViewPanoramaInitDemo> {
+  Uint8List? _bluePoint;
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -23,8 +35,8 @@ class StreetViewPanoramaInitDemo extends StatelessWidget {
                  * choice one of initPos or initPanoId
                  * do not feed param to both of them, or you should get assert error
                  */
-                //initPos: SAN_FRAN,
-                initPos: LatLng(25.0780892, 121.5753234),
+                initPos: SAN_FRAN,
+                //initPos: LatLng(25.0780892, 121.5753234),
                 //initPanoId: SANTORINI,
 
                 /**
@@ -83,6 +95,32 @@ class StreetViewPanoramaInitDemo extends StatelessWidget {
                  *  default setting is true
                  */
                 zoomGesturesEnabled: false,
+
+                /**
+                 *  iOS only
+                 *  Add marker to street view.
+                 */
+                markers: <Marker>[
+                  Marker(
+                    icon: _bluePoint == null
+                        ? BitmapDescriptor.defaultMarker
+                        : BitmapDescriptor.fromBytes(_bluePoint!),
+                    markerId: MarkerId("0"),
+                    position: SAN_FRAN,
+                    onTap: () {
+                      if (_bluePoint == null)
+                        DefaultAssetBundle.of(context)
+                            .load("assets/images/ic_dot.png")
+                            .then((data) {
+                          setState(() {
+                            _bluePoint = data.buffer.asUint8List();
+                          });
+                        });
+                      else
+                        setState(() => _bluePoint = null);
+                    },
+                  )
+                ].toSet(),
 
                 // Web only
                 //addressControl: false,