+++ flutter_google_street_view/ios/Classes/SwiftFlutterGoogleStreetViewPlugin.swift
@@ -3,7 +3,7 @@ import UIKit
 
 public class SwiftFlutterGoogleStreetViewPlugin: NSObject, FlutterPlugin {
   public static func register(with registrar: FlutterPluginRegistrar) {
-      let factory = FlutterGoogleStreetViewFactory(messenger: registrar.messenger())
+      let factory = FlutterGoogleStreetViewFactory(registrar: registrar)
             registrar.register(factory, withId: "my_street_view")
   }
 }