+++ flutter_google_street_view/lib/src/street_view_state.dart
@@ -13,4 +13,8 @@ class StreetViewState extends State<FlutterGoogleStreetView> {
   }
 
   void _onPlatformViewCreated(int id) async {}
+
+  void onMarkerTap(MarkerId markerId) {
+    throw UnimplementedError();
+  }
 }