+++ flutter_google_street_view/lib/src/web/street_view_state.dart
@@ -85,6 +85,10 @@ class StreetViewState extends State<FlutterGoogleStreetView> {
     _streetViewOptions = newOptions;
   }
 
+  void onMarkerTap(MarkerId markerId) {
+    throw UnimplementedError();
+  }
+
   void _onPlatformViewCreated(int id) async {
     final StreetViewController controller =
         await StreetViewController.init(id, this);